+++ src/app/components/layouts/column-container/two-columns/terra-two-columns-container.directive.ts
@@ -14,8 +14,8 @@ import {
 } from '@angular/router';
 import { Event } from '@angular/router/src/events';
 import { Observable } from 'rxjs/Observable';
-import { TwoColumnHelper } from '../../../helpers/two-column.helper';
-import { TerraTwoColumnsContainerComponent } from './two-columns/terra-two-columns-container.component';
+import { TwoColumnHelper } from '../../../../helpers/two-column.helper';
+import { TerraTwoColumnsContainerComponent } from './terra-two-columns-container.component';
 import { isNullOrUndefined } from 'util';
 import { Subscription } from 'rxjs/Subscription';
